+++ HISTORY.md
@@ -6,6 +6,10 @@
 - [ ] `BTR_String_freeToBuffer` function.
 - [ ] `BTR_String_copyToBuffer` function.
 - [x] `BTR_BHTable_put` now returns replaced item.
+- [x] `BTR_StringView_trimLeftTo` function.
+- [x] `BTR_StringView_trimLeftPast` function.
+- [x] `BTR_StringView_trimRightTo` function.
+- [x] `BTR_StringView_trimRightPast` function.
 ### Fixes
 - [ ] `BTR_String_append` (also prepend) fix for char.
 - [x] `BTR_StringView_findView` off-by-one error fix.